Mark as unmaintained
Add in a notice about the upcoming lack of intended maintenance on this repository. We'd like to make it clear to visitors to this repository about it's state. This process of marking it unmaintained only affects future updates. The Action will still continue work as it is in old workflows.

Closes #119

Once this lands, we'll also Archive this repository.

If you've got good recommendations for replacements, please let me know, we'll add them into the README.